refactor: use store cache for ProcessBaseline datastore
