+++ central/processbaseline/datastore/datastore.go
@@ -2,20 +2,16 @@ package datastore
 
 import (
 	"context"
-	"testing"
 
 	"github.com/stackrox/rox/central/globaldb"
 	"github.com/stackrox/rox/central/processbaseline/search"
 	"github.com/stackrox/rox/central/processbaseline/store"
-	pgStore "github.com/stackrox/rox/central/processbaseline/store/postgres"
 	"github.com/stackrox/rox/central/processbaselineresults/datastore"
 	processIndicatorDatastore "github.com/stackrox/rox/central/processindicator/datastore"
 	v1 "github.com/stackrox/rox/generated/api/v1"
 	"github.com/stackrox/rox/generated/storage"
 	"github.com/stackrox/rox/pkg/concurrency"
-	"github.com/stackrox/rox/pkg/postgres"
 	pkgSearch "github.com/stackrox/rox/pkg/search"
-	"github.com/stretchr/testify/require"
 )
 
 // DataStore wraps storage, and searcher for ProcessBaselines.
@@ -52,25 +48,4 @@ func New(storage store.Store, searcher search.Searcher, processBaselineResults d
 		processesDataStore:     processIndicators,
 	}
 	return d
-}
-
-// GetTestPostgresDataStore provides a datastore connected to postgres for testing purposes.
-func GetTestPostgresDataStore(t testing.TB, pool postgres.DB) (DataStore, error) {
-	dbStore := pgStore.New(pool)
-	store, err := pgStore.NewWithCache(dbStore)
-	require.NoError(t, err)
-	indexer := pgStore.NewIndexer(pool)
-	searcher, err := search.New(store, indexer)
-	if err != nil {
-		return nil, err
-	}
-	resultsStore, err := datastore.GetTestPostgresDataStore(t, pool)
-	if err != nil {
-		return nil, err
-	}
-	indicatorStore, err := processIndicatorDatastore.GetTestPostgresDataStore(t, pool)
-	if err != nil {
-		return nil, err
-	}
-	return New(store, searcher, resultsStore, indicatorStore), nil
 }
