+++ central/processbaseline/service/service_impl_test.go
@@ -102,10 +102,7 @@ func (suite *ProcessBaselineServiceTestSuite) SetupTest() {
 	pgtestbase := pgtest.ForT(suite.T())
 	suite.Require().NotNil(pgtestbase)
 	suite.pool = pgtestbase.DB
-	dbStore := postgresStore.New(suite.pool)
-	cache, err := postgresStore.NewWithCache(dbStore)
-	suite.NoError(err)
-	store := cache
+	store := postgresStore.New(suite.pool)
 	indexer := postgresStore.NewIndexer(suite.pool)
 
 	searcher, err := baselineSearch.New(store, indexer)
