+++ central/processbaseline/store/postgres/gen.go
@@ -1,3 +1,3 @@
 package postgres
 
-//go:generate pg-table-bindings-wrapper --type=storage.ProcessBaseline --search-category PROCESS_BASELINES
+//go:generate pg-table-bindings-wrapper --type=storage.ProcessBaseline --cached-store --search-category PROCESS_BASELINES
